fix: add git server + kind properties
also moved model function to the right source file
