+++ pkg/quickstarts/model.go
@@ -224,3 +224,31 @@ func (model *QuickstartModel) LoadQuickStarts(quickstarts *versionstream.QuickSt
 	}
 	return nil
 }
+
+func (model *QuickstartModel) convertToQuickStart(from *versionstream.QuickStart, to *Quickstart) error {
+	s := func(text string, override string) string {
+		if override != "" {
+			return override
+		}
+		return text
+	}
+	ss := func(texts []string, overrides []string) []string {
+		answer := append([]string{}, texts...)
+		for _, o := range overrides {
+			if util.StringArrayIndex(answer, o) < 0 {
+				answer = append(answer, o)
+			}
+		}
+		return answer
+	}
+
+	to.ID = s(to.ID, from.ID)
+	to.Owner = s(to.Owner, from.Owner)
+	to.Name = s(to.Name, from.Name)
+	to.Version = s(to.Version, from.Version)
+	to.DownloadZipURL = s(to.DownloadZipURL, from.DownloadZipURL)
+	to.Framework = s(to.Framework, from.Framework)
+	to.Language = s(to.Language, from.Language)
+	to.Tags = ss(to.Tags, from.Tags)
+	return nil
+}
